requirements.txt
ultralytics==8.0.230
opencv-python==4.8.1.78
fastapi==0.104.1
uvicorn==0.24.0
pillow>=11.0.0
numpy==1.26.2
python-multipart==0.0.6
pyyaml==6.0.1

reportlab==4.*     # better PDF reports
